Cleaned the code, added routes, some simple styling in the admin panel
