+++ client/components/admin.css
@@ -1,0 +1,13 @@
+#adminview button {
+  width: 40%;
+  background-color: blueviolet;
+  border-radius: 8px;
+  margin: 10px;
+  padding: 2%;
+  font-size: larger;
+  transition-duration: 0.4s;
+}
+
+#adminview button:hover {
+  background-color: white;
+}
